+++ x/inflation/keeper/epoch_info.go
@@ -38,3 +38,20 @@ func (k Keeper) SetEpochsPerPeriod(ctx sdk.Context, epochsPerPeriod int64) {
 	store := ctx.KVStore(k.storeKey)
 	store.Set(types.KeyPrefixEpochsPerPeriod, sdk.Uint64ToBigEndian(uint64(epochsPerPeriod)))
 }
+
+// GetSkippedEpochs gets the number of skipped epochs
+func (k Keeper) GetSkippedEpochs(ctx sdk.Context) uint64 {
+	store := ctx.KVStore(k.storeKey)
+	bz := store.Get(types.KeyPrefixSkippedEpochs)
+	if len(bz) == 0 {
+		return 0
+	}
+
+	return sdk.BigEndianToUint64(bz)
+}
+
+// SetSkippedEpochs stores the number of skipped epochs
+func (k Keeper) SetSkippedEpochs(ctx sdk.Context, skippedEpochs uint64) {
+	store := ctx.KVStore(k.storeKey)
+	store.Set(types.KeyPrefixSkippedEpochs, sdk.Uint64ToBigEndian(skippedEpochs))
+}