+++ x/inflation/keeper/epoch_info_test.go
@@ -81,3 +81,41 @@ func (suite *KeeperTestSuite) TestSetGetEpochsPerPeriod() {
 		})
 	}
 }
+
+func (suite *KeeperTestSuite) TestSetGetSkippedEpochs() {
+	defaultSkippedEpochs := types.DefaultGenesisState().SkippedEpochs
+	expSkippedepochs := uint64(20)
+
+	testCases := []struct {
+		name     string
+		malleate func()
+		ok       bool
+	}{
+		{
+			"default skipped epoch",
+			func() {},
+			false,
+		},
+		{
+			"skipped epoch set",
+			func() {
+				suite.app.InflationKeeper.SetSkippedEpochs(suite.ctx, expSkippedepochs)
+			},
+			true,
+		},
+	}
+	for _, tc := range testCases {
+		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
+			suite.SetupTest() // reset
+
+			tc.malleate()
+
+			epochsPerPeriod := suite.app.InflationKeeper.GetSkippedEpochs(suite.ctx)
+			if tc.ok {
+				suite.Require().Equal(expSkippedepochs, epochsPerPeriod, tc.name)
+			} else {
+				suite.Require().Equal(defaultSkippedEpochs, epochsPerPeriod, tc.name)
+			}
+		})
+	}
+}