+++ x/inflation/types/query.pb.go
@@ -191,6 +191,89 @@ func (m *QueryEpochMintProvisionResponse) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_QueryEpochMintProvisionResponse proto.InternalMessageInfo
 
+// QuerySkippedEpochsRequest is the request type for the Query/SkippedEpochs RPC method.
+type QuerySkippedEpochsRequest struct {
+}
+
+func (m *QuerySkippedEpochsRequest) Reset()         { *m = QuerySkippedEpochsRequest{} }
+func (m *QuerySkippedEpochsRequest) String() string { return proto.CompactTextString(m) }
+func (*QuerySkippedEpochsRequest) ProtoMessage()    {}
+func (*QuerySkippedEpochsRequest) Descriptor() ([]byte, []int) {
+	return fileDescriptor_91b9f1b5d47c7fd7, []int{4}
+}
+func (m *QuerySkippedEpochsRequest) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QuerySkippedEpochsRequest) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QuerySkippedEpochsRequest.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QuerySkippedEpochsRequest) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QuerySkippedEpochsRequest.Merge(m, src)
+}
+func (m *QuerySkippedEpochsRequest) XXX_Size() int {
+	return m.Size()
+}
+func (m *QuerySkippedEpochsRequest) XXX_DiscardUnknown() {
+	xxx_messageInfo_QuerySkippedEpochsRequest.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QuerySkippedEpochsRequest proto.InternalMessageInfo
+
+// QuerySkippedEpochsResponse is the response type for the Query/SkippedEpochs RPC method.
+type QuerySkippedEpochsResponse struct {
+	// number of epochs that the inflation module has been disabled.
+	SkippedEpochs uint64 `protobuf:"varint,1,opt,name=skipped_epochs,json=skippedEpochs,proto3" json:"skipped_epochs,omitempty"`
+}
+
+func (m *QuerySkippedEpochsResponse) Reset()         { *m = QuerySkippedEpochsResponse{} }
+func (m *QuerySkippedEpochsResponse) String() string { return proto.CompactTextString(m) }
+func (*QuerySkippedEpochsResponse) ProtoMessage()    {}
+func (*QuerySkippedEpochsResponse) Descriptor() ([]byte, []int) {
+	return fileDescriptor_91b9f1b5d47c7fd7, []int{5}
+}
+func (m *QuerySkippedEpochsResponse) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *QuerySkippedEpochsResponse) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_QuerySkippedEpochsResponse.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *QuerySkippedEpochsResponse) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_QuerySkippedEpochsResponse.Merge(m, src)
+}
+func (m *QuerySkippedEpochsResponse) XXX_Size() int {
+	return m.Size()
+}
+func (m *QuerySkippedEpochsResponse) XXX_DiscardUnknown() {
+	xxx_messageInfo_QuerySkippedEpochsResponse.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_QuerySkippedEpochsResponse proto.InternalMessageInfo
+
+func (m *QuerySkippedEpochsResponse) GetSkippedEpochs() uint64 {
+	if m != nil {
+		return m.SkippedEpochs
+	}
+	return 0
+}
+
 // QueryParamsRequest is the request type for the Query/Params RPC method.
 type QueryParamsRequest struct {
 }
@@ -199,7 +282,7 @@ func (m *QueryParamsRequest) Reset()         { *m = QueryParamsRequest{} }
 func (m *QueryParamsRequest) String() string { return proto.CompactTextString(m) }
 func (*QueryParamsRequest) ProtoMessage()    {}
 func (*QueryParamsRequest) Descriptor() ([]byte, []int) {
-	return fileDescriptor_91b9f1b5d47c7fd7, []int{4}
+	return fileDescriptor_91b9f1b5d47c7fd7, []int{6}
 }
 func (m *QueryParamsRequest) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -238,7 +321,7 @@ func (m *QueryParamsResponse) Reset()         { *m = QueryParamsResponse{} }
 func (m *QueryParamsResponse) String() string { return proto.CompactTextString(m) }
 func (*QueryParamsResponse) ProtoMessage()    {}
 func (*QueryParamsResponse) Descriptor() ([]byte, []int) {
-	return fileDescriptor_91b9f1b5d47c7fd7, []int{5}
+	return fileDescriptor_91b9f1b5d47c7fd7, []int{7}
 }
 func (m *QueryParamsResponse) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
@@ -279,6 +362,8 @@ func init() {
 	proto.RegisterType((*QueryPeriodResponse)(nil), "evmos.inflation.v1.QueryPeriodResponse")
 	proto.RegisterType((*QueryEpochMintProvisionRequest)(nil), "evmos.inflation.v1.QueryEpochMintProvisionRequest")
 	proto.RegisterType((*QueryEpochMintProvisionResponse)(nil), "evmos.inflation.v1.QueryEpochMintProvisionResponse")
+	proto.RegisterType((*QuerySkippedEpochsRequest)(nil), "evmos.inflation.v1.QuerySkippedEpochsRequest")
+	proto.RegisterType((*QuerySkippedEpochsResponse)(nil), "evmos.inflation.v1.QuerySkippedEpochsResponse")
 	proto.RegisterType((*QueryParamsRequest)(nil), "evmos.inflation.v1.QueryParamsRequest")
 	proto.RegisterType((*QueryParamsResponse)(nil), "evmos.inflation.v1.QueryParamsResponse")
 }
@@ -286,35 +371,39 @@ func init() {
 func init() { proto.RegisterFile("evmos/inflation/v1/query.proto", fileDescriptor_91b9f1b5d47c7fd7) }
 
 var fileDescriptor_91b9f1b5d47c7fd7 = []byte{
-	// 443 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x53, 0x31, 0x6f, 0xd4, 0x30,
-	0x14, 0x4e, 0xa0, 0x64, 0x30, 0x4c, 0xe6, 0x84, 0x50, 0x54, 0xf9, 0x4e, 0x19, 0x4a, 0x85, 0x74,
-	0x36, 0x4d, 0x17, 0xe6, 0x0a, 0xc4, 0x84, 0x28, 0x19, 0x59, 0x4a, 0x9a, 0x9a, 0x9c, 0x45, 0xe3,
-	0xe7, 0xc6, 0xbe, 0x88, 0x0e, 0x2c, 0xf0, 0x07, 0x90, 0xf8, 0x21, 0x8c, 0xfc, 0x85, 0x8e, 0x95,
-	0x58, 0x10, 0xc3, 0x09, 0xdd, 0xf1, 0x43, 0x50, 0x1c, 0x07, 0xee, 0x94, 0xdc, 0x9d, 0x3a, 0x25,
-	0x79, 0xdf, 0xe7, 0xf7, 0xbe, 0xf7, 0x7d, 0x0e, 0x22, 0xbc, 0x2a, 0x40, 0x33, 0x21, 0xdf, 0x9d,
-	0xa7, 0x46, 0x80, 0x64, 0xd5, 0x01, 0xbb, 0x98, 0xf2, 0xf2, 0x92, 0xaa, 0x12, 0x0c, 0x60, 0x6c,
-	0x71, 0xfa, 0x0f, 0xa7, 0xd5, 0x41, 0x38, 0xc8, 0x21, 0x07, 0x0b, 0xb3, 0xfa, 0xad, 0x61, 0x86,
-	0xbb, 0x39, 0x40, 0x7e, 0xce, 0x59, 0xaa, 0x04, 0x4b, 0xa5, 0x04, 0x63, 0xf9, 0xda, 0xa1, 0xa3,
-	0x9e, 0x39, 0x39, 0x97, 0x5c, 0x0b, 0xc7, 0x88, 0x06, 0x08, 0xbf, 0xae, 0x07, 0x1f, 0xf3, 0x52,
-	0xc0, 0x59, 0xc2, 0x2f, 0xa6, 0x5c, 0x9b, 0x68, 0x8c, 0xee, 0xaf, 0x54, 0xb5, 0x02, 0xa9, 0x39,
-	0x7e, 0x80, 0x02, 0x65, 0x2b, 0x0f, 0xfd, 0x91, 0xbf, 0xbf, 0x93, 0xb8, 0xaf, 0x68, 0x84, 0x88,
-	0xa5, 0x3f, 0x57, 0x90, 0x4d, 0x5e, 0x0a, 0x69, 0x8e, 0x4b, 0xa8, 0x84, 0x16, 0x20, 0xdb, 0x86,
-	0x9f, 0x7d, 0x34, 0x5c, 0x4b, 0x71, 0xdd, 0xdf, 0xa2, 0x01, 0xaf, 0xd1, 0x93, 0x42, 0x48, 0x73,
-	0xa2, 0x5a, 0xdc, 0xce, 0xba, 0x77, 0x44, 0xaf, 0x66, 0x43, 0xef, 0xd7, 0x6c, 0xb8, 0x97, 0x0b,
-	0x33, 0x99, 0x9e, 0xd2, 0x0c, 0x0a, 0x96, 0x81, 0xae, 0xd7, 0x6b, 0x1e, 0x63, 0x7d, 0xf6, 0x9e,
-	0x99, 0x4b, 0xc5, 0x35, 0x7d, 0xc6, 0xb3, 0x04, 0xf3, 0xce, 0xa4, 0xff, 0xcb, 0xa6, 0x65, 0x5a,
-	0xe8, 0x56, 0xdb, 0xab, 0x76, 0x59, 0x57, 0x75, 0x72, 0x9e, 0xa2, 0x40, 0xd9, 0x8a, 0x15, 0x70,
-	0x37, 0x0e, 0x69, 0x37, 0x14, 0xda, 0x9c, 0x39, 0xda, 0xa9, 0xc5, 0x25, 0x8e, 0x1f, 0x7f, 0xbf,
-	0x8d, 0xee, 0xd8, 0x8e, 0xf8, 0x23, 0x0a, 0x1a, 0x0b, 0xf1, 0x5e, 0xdf, 0xe9, 0xae, 0xf3, 0xe1,
-	0xa3, 0xad, 0xbc, 0x46, 0x5e, 0x14, 0x7d, 0xfa, 0xf1, 0xe7, 0xeb, 0xad, 0x5d, 0x1c, 0xb2, 0x9e,
-	0x8c, 0x9b, 0x5c, 0xf0, 0x37, 0x1f, 0xe1, 0xae, 0xe1, 0x38, 0x5e, 0x3b, 0x63, 0x6d, 0x80, 0xe1,
-	0xe1, 0x8d, 0xce, 0x38, 0x8d, 0x4f, 0xac, 0xc6, 0xc7, 0x78, 0xbf, 0x4f, 0x63, 0x5f, 0xd6, 0xd6,
-	0x30, 0x6b, 0xe2, 0x26, 0xc3, 0x96, 0xd3, 0xdb, 0x64, 0xd8, 0x4a, 0x9e, 0x5b, 0x0c, 0x6b, 0x72,
-	0x7c, 0x71, 0x35, 0x27, 0xfe, 0xf5, 0x9c, 0xf8, 0xbf, 0xe7, 0xc4, 0xff, 0xb2, 0x20, 0xde, 0xf5,
-	0x82, 0x78, 0x3f, 0x17, 0xc4, 0x7b, 0x33, 0x5e, 0xba, 0x76, 0x66, 0x92, 0x96, 0x5a, 0x68, 0xd7,
-	0xa7, 0x8a, 0xd9, 0x87, 0xa5, 0x66, 0xf6, 0x06, 0x9e, 0x06, 0xf6, 0xef, 0x3a, 0xfc, 0x1b, 0x00,
-	0x00, 0xff, 0xff, 0x02, 0xd9, 0x9f, 0xce, 0xe9, 0x03, 0x00, 0x00,
+	// 509 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x94, 0x54, 0x31, 0x6f, 0x13, 0x31,
+	0x18, 0xcd, 0xa1, 0x90, 0xc1, 0x50, 0x06, 0x13, 0x21, 0x38, 0xaa, 0x4b, 0x74, 0x82, 0x52, 0x55,
+	0x8a, 0x4d, 0xd3, 0x85, 0xb9, 0x80, 0x98, 0x10, 0x25, 0x6c, 0x2c, 0xe1, 0x9a, 0x98, 0x8b, 0xd5,
+	0xc6, 0x76, 0xef, 0x73, 0x4e, 0x74, 0x60, 0x81, 0x3f, 0x80, 0xc4, 0xc6, 0x9f, 0xe0, 0x6f, 0x74,
+	0xac, 0x04, 0x03, 0x62, 0xa8, 0x50, 0xc2, 0x0f, 0x41, 0xf7, 0x9d, 0x53, 0x12, 0xc5, 0xd7, 0xc2,
+	0x94, 0xe4, 0x7b, 0xcf, 0xef, 0x7b, 0xe7, 0xf7, 0x2e, 0x24, 0x12, 0xf9, 0x58, 0x03, 0x97, 0xea,
+	0xed, 0x61, 0x62, 0xa5, 0x56, 0x3c, 0xdf, 0xe6, 0x47, 0x13, 0x91, 0x1d, 0x33, 0x93, 0x69, 0xab,
+	0x29, 0x45, 0x9c, 0x9d, 0xe3, 0x2c, 0xdf, 0x0e, 0x9b, 0xa9, 0x4e, 0x35, 0xc2, 0xbc, 0xf8, 0x56,
+	0x32, 0xc3, 0xf5, 0x54, 0xeb, 0xf4, 0x50, 0xf0, 0xc4, 0x48, 0x9e, 0x28, 0xa5, 0x2d, 0xf2, 0xc1,
+	0xa1, 0x6d, 0xcf, 0x9e, 0x54, 0x28, 0x01, 0xd2, 0x31, 0xe2, 0x26, 0xa1, 0x2f, 0x8b, 0xc5, 0x7b,
+	0x22, 0x93, 0x7a, 0xd8, 0x13, 0x47, 0x13, 0x01, 0x36, 0xee, 0x90, 0x9b, 0x4b, 0x53, 0x30, 0x5a,
+	0x81, 0xa0, 0xb7, 0x48, 0xc3, 0xe0, 0xe4, 0x76, 0xd0, 0x0e, 0x36, 0xeb, 0x3d, 0xf7, 0x2b, 0x6e,
+	0x93, 0x08, 0xe9, 0x4f, 0x8d, 0x1e, 0x8c, 0x9e, 0x4b, 0x65, 0xf7, 0x32, 0x9d, 0x4b, 0x90, 0x5a,
+	0xcd, 0x05, 0x3f, 0x06, 0xa4, 0x55, 0x49, 0x71, 0xea, 0x6f, 0x48, 0x53, 0x14, 0x68, 0x7f, 0x2c,
+	0x95, 0xed, 0x9b, 0x39, 0x8e, 0xbb, 0xae, 0xef, 0xb2, 0x93, 0xb3, 0x56, 0xed, 0xe7, 0x59, 0x6b,
+	0x23, 0x95, 0x76, 0x34, 0xd9, 0x67, 0x03, 0x3d, 0xe6, 0x03, 0x0d, 0xc5, 0xe3, 0x95, 0x1f, 0x1d,
+	0x18, 0x1e, 0x70, 0x7b, 0x6c, 0x04, 0xb0, 0x27, 0x62, 0xd0, 0xa3, 0x62, 0x65, 0x53, 0x7c, 0x97,
+	0xdc, 0x41, 0x13, 0xaf, 0x0e, 0xa4, 0x31, 0x62, 0x88, 0x5e, 0x60, 0x6e, 0xf1, 0x31, 0x09, 0x7d,
+	0xa0, 0x33, 0x77, 0x9f, 0xdc, 0x80, 0x12, 0xe8, 0xa3, 0x30, 0xb8, 0x2b, 0x58, 0x83, 0x45, 0xfa,
+	0xdf, 0xeb, 0x4c, 0xb2, 0x64, 0x7c, 0x2e, 0xfd, 0x62, 0x7e, 0x9d, 0x6e, 0xea, 0x34, 0x1f, 0x91,
+	0x86, 0xc1, 0x09, 0x6a, 0x5d, 0xeb, 0x86, 0x6c, 0x35, 0x76, 0x56, 0x9e, 0xd9, 0xad, 0x17, 0x8f,
+	0xdf, 0x73, 0xfc, 0xee, 0xf7, 0x3a, 0xb9, 0x8a, 0x8a, 0xf4, 0x3d, 0x69, 0x94, 0x21, 0xd1, 0x0d,
+	0xdf, 0xe9, 0xd5, 0x6c, 0xc3, 0x07, 0x97, 0xf2, 0x4a, 0x7b, 0x71, 0xfc, 0xe1, 0xdb, 0xef, 0xcf,
+	0x57, 0xd6, 0x69, 0xc8, 0x3d, 0x2d, 0x2a, 0x93, 0xa7, 0x5f, 0x03, 0x42, 0x57, 0x23, 0xa5, 0xdd,
+	0xca, 0x1d, 0x95, 0x15, 0x09, 0x77, 0xfe, 0xeb, 0x8c, 0xf3, 0xf8, 0x10, 0x3d, 0x6e, 0xd1, 0x4d,
+	0x9f, 0x47, 0x5f, 0x9b, 0xe8, 0x97, 0x80, 0xac, 0x2d, 0x45, 0x4c, 0x3b, 0x95, 0x8b, 0x7d, 0x3d,
+	0x09, 0xd9, 0xbf, 0xd2, 0x9d, 0xc5, 0x2d, 0xb4, 0x78, 0x8f, 0xc6, 0x3e, 0x8b, 0xcb, 0x9d, 0xc2,
+	0x34, 0x31, 0xe1, 0x8b, 0xd2, 0x5c, 0xac, 0xd6, 0x45, 0x69, 0x2e, 0x95, 0xed, 0x92, 0x34, 0xcb,
+	0x92, 0x3d, 0x3b, 0x99, 0x46, 0xc1, 0xe9, 0x34, 0x0a, 0x7e, 0x4d, 0xa3, 0xe0, 0xd3, 0x2c, 0xaa,
+	0x9d, 0xce, 0xa2, 0xda, 0x8f, 0x59, 0x54, 0x7b, 0xdd, 0x59, 0x78, 0xeb, 0xec, 0x28, 0xc9, 0x40,
+	0x82, 0xd3, 0xc9, 0xbb, 0xfc, 0xdd, 0x82, 0x18, 0xbe, 0x80, 0xfb, 0x0d, 0xfc, 0x73, 0xd9, 0xf9,
+	0x13, 0x00, 0x00, 0xff, 0xff, 0xd9, 0x37, 0xcd, 0x6f, 0xe8, 0x04, 0x00, 0x00,
 }
 
 // Reference imports to suppress errors if they are not otherwise used.
@@ -333,6 +422,8 @@ type QueryClient interface {
 	Period(ctx context.Context, in *QueryPeriodRequest, opts ...grpc.CallOption) (*QueryPeriodResponse, error)
 	// EpochMintProvision retrieves current minting epoch provision value.
 	EpochMintProvision(ctx context.Context, in *QueryEpochMintProvisionRequest, opts ...grpc.CallOption) (*QueryEpochMintProvisionResponse, error)
+	// SkippedEpochs retrieves the total number of skipped epochs.
+	SkippedEpochs(ctx context.Context, in *QuerySkippedEpochsRequest, opts ...grpc.CallOption) (*QuerySkippedEpochsResponse, error)
 	// Params retrieves the total set of minting parameters.
 	Params(ctx context.Context, in *QueryParamsRequest, opts ...grpc.CallOption) (*QueryParamsResponse, error)
 }
@@ -363,6 +454,15 @@ func (c *queryClient) EpochMintProvision(ctx context.Context, in *QueryEpochMint
 	return out, nil
 }
 
+func (c *queryClient) SkippedEpochs(ctx context.Context, in *QuerySkippedEpochsRequest, opts ...grpc.CallOption) (*QuerySkippedEpochsResponse, error) {
+	out := new(QuerySkippedEpochsResponse)
+	err := c.cc.Invoke(ctx, "/evmos.inflation.v1.Query/SkippedEpochs", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 func (c *queryClient) Params(ctx context.Context, in *QueryParamsRequest, opts ...grpc.CallOption) (*QueryParamsResponse, error) {
 	out := new(QueryParamsResponse)
 	err := c.cc.Invoke(ctx, "/evmos.inflation.v1.Query/Params", in, out, opts...)
@@ -378,6 +478,8 @@ type QueryServer interface {
 	Period(context.Context, *QueryPeriodRequest) (*QueryPeriodResponse, error)
 	// EpochMintProvision retrieves current minting epoch provision value.
 	EpochMintProvision(context.Context, *QueryEpochMintProvisionRequest) (*QueryEpochMintProvisionResponse, error)
+	// SkippedEpochs retrieves the total number of skipped epochs.
+	SkippedEpochs(context.Context, *QuerySkippedEpochsRequest) (*QuerySkippedEpochsResponse, error)
 	// Params retrieves the total set of minting parameters.
 	Params(context.Context, *QueryParamsRequest) (*QueryParamsResponse, error)
 }
@@ -391,6 +493,9 @@ func (*UnimplementedQueryServer) Period(ctx context.Context, req *QueryPeriodReq
 }
 func (*UnimplementedQueryServer) EpochMintProvision(ctx context.Context, req *QueryEpochMintProvisionRequest) (*QueryEpochMintProvisionResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method EpochMintProvision not implemented")
+}
+func (*UnimplementedQueryServer) SkippedEpochs(ctx context.Context, req *QuerySkippedEpochsRequest) (*QuerySkippedEpochsResponse, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method SkippedEpochs not implemented")
 }
 func (*UnimplementedQueryServer) Params(ctx context.Context, req *QueryParamsRequest) (*QueryParamsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Params not implemented")
@@ -436,6 +541,24 @@ func _Query_EpochMintProvision_Handler(srv interface{}, ctx context.Context, dec
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Query_SkippedEpochs_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(QuerySkippedEpochsRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(QueryServer).SkippedEpochs(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/evmos.inflation.v1.Query/SkippedEpochs",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(QueryServer).SkippedEpochs(ctx, req.(*QuerySkippedEpochsRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _Query_Params_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(QueryParamsRequest)
 	if err := dec(in); err != nil {
@@ -465,6 +588,10 @@ var _Query_serviceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "EpochMintProvision",
 			Handler:    _Query_EpochMintProvision_Handler,
+		},
+		{
+			MethodName: "SkippedEpochs",
+			Handler:    _Query_SkippedEpochs_Handler,
 		},
 		{
 			MethodName: "Params",
@@ -582,6 +709,57 @@ func (m *QueryEpochMintProvisionResponse) MarshalToSizedBuffer(dAtA []byte) (int
 	return len(dAtA) - i, nil
 }
 
+func (m *QuerySkippedEpochsRequest) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QuerySkippedEpochsRequest) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QuerySkippedEpochsRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	return len(dAtA) - i, nil
+}
+
+func (m *QuerySkippedEpochsResponse) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *QuerySkippedEpochsResponse) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *QuerySkippedEpochsResponse) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if m.SkippedEpochs != 0 {
+		i = encodeVarintQuery(dAtA, i, uint64(m.SkippedEpochs))
+		i--
+		dAtA[i] = 0x8
+	}
+	return len(dAtA) - i, nil
+}
+
 func (m *QueryParamsRequest) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
@@ -687,6 +865,27 @@ func (m *QueryEpochMintProvisionResponse) Size() (n int) {
 	_ = l
 	l = m.EpochMintProvision.Size()
 	n += 1 + l + sovQuery(uint64(l))
+	return n
+}
+
+func (m *QuerySkippedEpochsRequest) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	return n
+}
+
+func (m *QuerySkippedEpochsResponse) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	if m.SkippedEpochs != 0 {
+		n += 1 + sovQuery(uint64(m.SkippedEpochs))
+	}
 	return n
 }
 
@@ -947,6 +1146,125 @@ func (m *QueryEpochMintProvisionResponse) Unmarshal(dAtA []byte) error {
 				return err
 			}
 			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQuery(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *QuerySkippedEpochsRequest) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQuery
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QuerySkippedEpochsRequest: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QuerySkippedEpochsRequest: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		default:
+			iNdEx = preIndex
+			skippy, err := skipQuery(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthQuery
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *QuerySkippedEpochsResponse) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowQuery
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: QuerySkippedEpochsResponse: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: QuerySkippedEpochsResponse: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field SkippedEpochs", wireType)
+			}
+			m.SkippedEpochs = 0
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowQuery
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				m.SkippedEpochs |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
 		default:
 			iNdEx = preIndex
 			skippy, err := skipQuery(dAtA[iNdEx:])